load the environment variables from .env in the unit test so we can use the actual table.
Note this is normally a bad idea: this should be an integration test. This is the only change I'm making to the test, though.
Also note that because this should be an integration test, it can't run successfully twice without manually removing the added entry because the entry will no longer be unique when adding it a second time.